moved bellatrix epoch and version to base spec config
In both of these, (bellatrix and altair) we don't really have  enough context to determine that values should be defaulted. So far it's been fine, but we'll likely need a better solution once we get closer to fulu, or potentially in the G fork.
