+++ eth-reference-tests/src/referenceTest/java/tech/pegasys/teku/reference/altair/fork/TransitionTestExecutor.java
@@ -55,31 +55,29 @@ public class TransitionTestExecutor implements TestExecutor {
               switch (milestone) {
                 case ALTAIR -> builder.altairForkEpoch(forkEpoch);
                 case BELLATRIX ->
-                    builder
-                        .altairForkEpoch(UInt64.ZERO)
-                        .bellatrixBuilder(b -> b.bellatrixForkEpoch(forkEpoch));
+                    builder.altairForkEpoch(UInt64.ZERO).bellatrixForkEpoch(forkEpoch);
                 case CAPELLA ->
                     builder
                         .altairForkEpoch(UInt64.ZERO)
-                        .bellatrixBuilder(b -> b.bellatrixForkEpoch(UInt64.ZERO))
+                        .bellatrixForkEpoch(UInt64.ZERO)
                         .capellaBuilder(c -> c.capellaForkEpoch(forkEpoch));
                 case DENEB ->
                     builder
                         .altairForkEpoch(UInt64.ZERO)
-                        .bellatrixBuilder(b -> b.bellatrixForkEpoch(UInt64.ZERO))
+                        .bellatrixForkEpoch(UInt64.ZERO)
                         .capellaBuilder(c -> c.capellaForkEpoch(UInt64.ZERO))
                         .denebBuilder(d -> d.denebForkEpoch(forkEpoch));
                 case ELECTRA ->
                     builder
                         .altairForkEpoch(UInt64.ZERO)
-                        .bellatrixBuilder(b -> b.bellatrixForkEpoch(UInt64.ZERO))
+                        .bellatrixForkEpoch(UInt64.ZERO)
                         .capellaBuilder(c -> c.capellaForkEpoch(UInt64.ZERO))
                         .denebBuilder(d -> d.denebForkEpoch(UInt64.ZERO))
                         .electraBuilder(e -> e.electraForkEpoch(forkEpoch));
                 case FULU ->
                     builder
                         .altairForkEpoch(UInt64.ZERO)
-                        .bellatrixBuilder(b -> b.bellatrixForkEpoch(UInt64.ZERO))
+                        .bellatrixForkEpoch(UInt64.ZERO)
                         .capellaBuilder(c -> c.capellaForkEpoch(UInt64.ZERO))
                         .denebBuilder(d -> d.denebForkEpoch(UInt64.ZERO))
                         .electraBuilder(e -> e.electraForkEpoch(UInt64.ZERO))
