+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/config/SpecConfigPhase0.java
@@ -125,6 +125,10 @@ public class SpecConfigPhase0 implements SpecConfig {
   private final Bytes4 altairForkVersion;
   private final UInt64 altairForkEpoch;
 
+  // bellatrix fork
+  private final Bytes4 bellatrixForkVersion;
+  private final UInt64 bellatrixForkEpoch;
+
   public SpecConfigPhase0(
       final Map<String, Object> rawConfig,
       final UInt64 eth1FollowDistance,
@@ -194,7 +198,9 @@ public class SpecConfigPhase0 implements SpecConfig {
       final int reorgParentWeightThreshold,
       final UInt64 maxPerEpochActivationExitChurnLimit,
       final Bytes4 altairForkVersion,
-      final UInt64 altairForkEpoch) {
+      final UInt64 altairForkEpoch,
+      final Bytes4 bellatrixForkVersion,
+      final UInt64 bellatrixForkEpoch) {
     this.rawConfig = rawConfig;
     this.eth1FollowDistance = eth1FollowDistance;
     this.maxCommitteesPerSlot = maxCommitteesPerSlot;
@@ -265,6 +271,8 @@ public class SpecConfigPhase0 implements SpecConfig {
     this.maxPerEpochActivationExitChurnLimit = maxPerEpochActivationExitChurnLimit;
     this.altairForkVersion = altairForkVersion;
     this.altairForkEpoch = altairForkEpoch;
+    this.bellatrixForkVersion = bellatrixForkVersion;
+    this.bellatrixForkEpoch = bellatrixForkEpoch;
   }
 
   @Override
@@ -400,6 +408,16 @@ public class SpecConfigPhase0 implements SpecConfig {
   @Override
   public UInt64 getAltairForkEpoch() {
     return altairForkEpoch;
+  }
+
+  @Override
+  public Bytes4 getBellatrixForkVersion() {
+    return bellatrixForkVersion;
+  }
+
+  @Override
+  public UInt64 getBellatrixForkEpoch() {
+    return bellatrixForkEpoch;
   }
 
   @Override
@@ -707,6 +725,8 @@ public class SpecConfigPhase0 implements SpecConfig {
         && Objects.equals(genesisForkVersion, that.genesisForkVersion)
         && Objects.equals(altairForkVersion, that.altairForkVersion)
         && Objects.equals(altairForkEpoch, that.altairForkEpoch)
+        && Objects.equals(bellatrixForkVersion, that.bellatrixForkVersion)
+        && Objects.equals(bellatrixForkEpoch, that.bellatrixForkEpoch)
         && Objects.equals(genesisDelay, that.genesisDelay)
         && Objects.equals(minEpochsToInactivityPenalty, that.minEpochsToInactivityPenalty)
         && Objects.equals(shardCommitteePeriod, that.shardCommitteePeriod)
@@ -783,6 +803,8 @@ public class SpecConfigPhase0 implements SpecConfig {
         attestationSubnetExtraBits,
         altairForkVersion,
         altairForkEpoch,
+        bellatrixForkVersion,
+        bellatrixForkEpoch,
         attestationSubnetPrefixBits);
   }
 }
