+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/config/builder/SpecConfigBuilder.java
@@ -137,6 +137,8 @@ public class SpecConfigBuilder {
   // altair fork information
   private Bytes4 altairForkVersion;
   private UInt64 altairForkEpoch;
+  private Bytes4 bellatrixForkVersion;
+  private UInt64 bellatrixForkEpoch;
 
   private UInt64 maxPerEpochActivationExitChurnLimit = UInt64.valueOf(256000000000L);
   private final BuilderChain<SpecConfig, SpecConfigFulu> builderChain =
@@ -174,6 +176,15 @@ public class SpecConfigBuilder {
     if (altairForkVersion == null) {
       altairForkVersion = SpecBuilderUtil.PLACEHOLDER_FORK_VERSION;
       rawConfig.put("ALTAIR_FORK_VERSION", altairForkVersion);
+    }
+    if (bellatrixForkEpoch == null) {
+      bellatrixForkEpoch = SpecConfig.FAR_FUTURE_EPOCH;
+      bellatrixForkVersion = SpecBuilderUtil.PLACEHOLDER_FORK_VERSION;
+    }
+    rawConfig.put("BELLATRIX_FORK_EPOCH", bellatrixForkEpoch);
+    if (bellatrixForkVersion == null) {
+      bellatrixForkVersion = SpecBuilderUtil.PLACEHOLDER_FORK_VERSION;
+      rawConfig.put("BELLATRIX_FORK_VERSION", bellatrixForkVersion);
     }
     validate();
     final SpecConfigAndParent<SpecConfig> config =
@@ -247,7 +258,9 @@ public class SpecConfigBuilder {
                 reorgParentWeightThreshold,
                 maxPerEpochActivationExitChurnLimit,
                 altairForkVersion,
-                altairForkEpoch));
+                altairForkEpoch,
+                bellatrixForkVersion,
+                bellatrixForkEpoch));
 
     return builderChain.build(config);
   }
@@ -321,6 +334,8 @@ public class SpecConfigBuilder {
     constants.put("reorgParentWeightThreshold", reorgParentWeightThreshold);
     constants.put("altairForkEpoch", altairForkEpoch);
     constants.put("altairForkVersion", altairForkVersion);
+    constants.put("bellatrixForkEpoch", bellatrixForkEpoch);
+    constants.put("bellatrixForkVersion", bellatrixForkVersion);
     return constants;
   }
 
@@ -468,6 +483,18 @@ public class SpecConfigBuilder {
   public SpecConfigBuilder altairForkEpoch(final UInt64 altairForkEpoch) {
     checkNotNull(altairForkEpoch);
     this.altairForkEpoch = altairForkEpoch;
+    return this;
+  }
+
+  public SpecConfigBuilder bellatrixForkVersion(final Bytes4 bellatrixForkVersion) {
+    checkNotNull(bellatrixForkVersion);
+    this.bellatrixForkVersion = bellatrixForkVersion;
+    return this;
+  }
+
+  public SpecConfigBuilder bellatrixForkEpoch(final UInt64 bellatrixForkEpoch) {
+    checkNotNull(bellatrixForkEpoch);
+    this.bellatrixForkEpoch = bellatrixForkEpoch;
     return this;
   }
