+++ ethereum/spec/src/test/java/tech/pegasys/teku/spec/SpecFactoryTest.java
@@ -48,31 +48,29 @@ class SpecFactoryTest {
                 case PHASE0 -> LOG.info("PHASE0");
                 case ALTAIR -> builder.altairForkEpoch(forkEpoch);
                 case BELLATRIX ->
-                    builder
-                        .altairForkEpoch(UInt64.ZERO)
-                        .bellatrixBuilder(b -> b.bellatrixForkEpoch(forkEpoch));
+                    builder.altairForkEpoch(UInt64.ZERO).bellatrixForkEpoch(forkEpoch);
                 case CAPELLA ->
                     builder
                         .altairForkEpoch(UInt64.ZERO)
-                        .bellatrixBuilder(b -> b.bellatrixForkEpoch(UInt64.ZERO))
+                        .bellatrixForkEpoch(UInt64.ZERO)
                         .capellaBuilder(c -> c.capellaForkEpoch(forkEpoch));
                 case DENEB ->
                     builder
                         .altairForkEpoch(UInt64.ZERO)
-                        .bellatrixBuilder(b -> b.bellatrixForkEpoch(UInt64.ZERO))
+                        .bellatrixForkEpoch(UInt64.ZERO)
                         .capellaBuilder(c -> c.capellaForkEpoch(UInt64.ZERO))
                         .denebBuilder(d -> d.denebForkEpoch(forkEpoch));
                 case ELECTRA ->
                     builder
                         .altairForkEpoch(UInt64.ZERO)
-                        .bellatrixBuilder(b -> b.bellatrixForkEpoch(UInt64.ZERO))
+                        .bellatrixForkEpoch(UInt64.ZERO)
                         .capellaBuilder(c -> c.capellaForkEpoch(UInt64.ZERO))
                         .denebBuilder(d -> d.denebForkEpoch(UInt64.ZERO))
                         .electraBuilder(e -> e.electraForkEpoch(forkEpoch));
                 case FULU ->
                     builder
                         .altairForkEpoch(UInt64.ZERO)
-                        .bellatrixBuilder(b -> b.bellatrixForkEpoch(UInt64.ZERO))
+                        .bellatrixForkEpoch(UInt64.ZERO)
                         .capellaBuilder(c -> c.capellaForkEpoch(UInt64.ZERO))
                         .denebBuilder(d -> d.denebForkEpoch(UInt64.ZERO))
                         .electraBuilder(e -> e.electraForkEpoch(UInt64.ZERO))
