+++ networking/eth2/src/test/java/tech/pegasys/teku/networking/eth2/ActiveEth2P2PNetworkTest.java
@@ -66,7 +66,7 @@ public class ActiveEth2P2PNetworkTest {
       TestSpecFactory.createMinimalFulu(
           b ->
               b.altairForkEpoch(altairForkEpoch)
-                  .bellatrixBuilder(bb -> bb.bellatrixForkEpoch(UInt64.valueOf(3)))
+                  .bellatrixForkEpoch(UInt64.valueOf(3))
                   .capellaBuilder(cb -> cb.capellaForkEpoch(UInt64.valueOf(4)))
                   .denebBuilder(db -> db.denebForkEpoch(UInt64.valueOf(5)))
                   .electraBuilder(eb -> eb.electraForkEpoch(UInt64.valueOf(6)))
